+++ src/components/ids-locale/ids-locale-data.ts
@@ -1,9 +1,0 @@
-/** Static Holder of locale information */
-abstract class IdsLocaleData {
-  static loadedLocales = new Map();
-
-  // The lang files that have been loaded up
-  static loadedLanguages = new Map();
-}
-
-export default IdsLocaleData;