+++ test/ids-radio/ids-radio-group-func-test.ts
@@ -4,8 +4,8 @@
 import IdsRadio from '../../src/components/ids-radio/ids-radio';
 import IdsRadioGroup from '../../src/components/ids-radio/ids-radio-group';
 import IdsContainer from '../../src/components/ids-container/ids-container';
-import IdsLocaleData from '../../src/components/ids-locale/ids-locale-data';
 import { messages as deMessages } from '../../src/components/ids-locale/data/de-messages';
+import IdsGlobal from '../../src/components/ids-global/ids-global';
 
 jest.useFakeTimers();
 
@@ -15,7 +15,7 @@ describe('IdsRadioGroup Component', () => {
 
   beforeEach(async () => {
     container = new IdsContainer();
-    IdsLocaleData.loadedLanguages.set('de', deMessages);
+    IdsGlobal.getLocale().loadedLanguages.set('de', deMessages);
 
     const rb1 = new IdsRadio();
     const rb2 = new IdsRadio();
@@ -392,7 +392,7 @@ describe('IdsRadioGroup Component', () => {
   });
 
   it('can change language from the container', async () => {
-    await container.setLanguage('de');
+    await IdsGlobal.getLocale().setLanguage('de');
     expect(rg.getAttribute('language')).toEqual('de');
   });
 });